fix n0_known
I suspect n0_known does not work as intended when multiple wavelengths are used in a calculation, but I have not attempted to fix this, because L. Bolla probably has a clear idea about how he would like to implement this. A fix could be included here or in tm.solve, where RefractiveIndex could be called with one wavelength at a time as argument inside the for w in wls loop.